@@ -123,7 +123,7 @@
     "react-native": "local-cli/wrong-react-native.js"
   },
   "peerDependencies": {
-    "react": "15.0.3-alpha.1"
+    "react": "15.1.0-alpha.1"
   },
   "dependencies": {
     "absolute-path": "^0.0.0",
@@ -186,7 +186,7 @@
     "flow-bin": "0.24.0",
     "jest-cli": "11.0.2",
     "portfinder": "0.4.0",
-    "react": "15.0.3-alpha.1",
+    "react": "^15.1.0-alpha.1",
     "shelljs": "0.6.0"
   }
 }
