@@ -18,6 +18,7 @@ var Prism = require('Prism');
 var React = require('React');
 var Site = require('Site');
 var slugify = require('slugify');
+var Metadata = require('Metadata');
 
 var styleReferencePattern = /^[^.]+\.propTypes\.style$/;
 
@@ -520,11 +521,15 @@ var Modal = React.createClass({
 
 var Autodocs = React.createClass({
   childContextTypes: {
-    permalink: React.PropTypes.string
+    permalink: React.PropTypes.string,
+    version: React.PropTypes.string
   },
 
   getChildContext: function() {
-    return {permalink: this.props.metadata.permalink};
+    return {
+      permalink: this.props.metadata.permalink,
+      version: Metadata.config.RN_VERSION || 'next'
+    };
   },
 
   renderFullDescription: function(docs) {
