@@ -47,14 +47,16 @@ var PickerAndroid = React.createClass({
     enabled: ReactPropTypes.bool,
     mode: ReactPropTypes.oneOf(['dialog', 'dropdown']),
     onValueChange: ReactPropTypes.func,
-    onFocus: ReactPropTypes.func,
-    onDismissed: ReactPropTypes.func,
     prompt: ReactPropTypes.string,
     testID: ReactPropTypes.string,
   },
 
   getInitialState: function() {
-    return this._stateFromProps(this.props);
+    var state = this._stateFromProps(this.props);
+    return {
+      ...state,
+      initialSelectedIndex: state.selectedIndex,
+    };
   },
 
   componentWillReceiveProps: function(nextProps) {
@@ -63,7 +65,7 @@ var PickerAndroid = React.createClass({
 
   // Translate prop and children into stuff that the native picker understands.
   _stateFromProps: function(props) {
-    var selectedIndex = -1;
+    var selectedIndex = 0;
     let items = ReactChildren.map(props.children, (child, index) => {
       if (child.props.value === props.selectedValue) {
         selectedIndex = index;
@@ -87,23 +89,14 @@ var PickerAndroid = React.createClass({
       enabled: this.props.enabled,
       items: this.state.items,
       mode: this.props.mode,
-      onChange: this._onFocusChange,
       onSelect: this._onChange,
       prompt: this.props.prompt,
-      selected: this.state.selectedIndex,
+      selected: this.state.initialSelectedIndex,
       testID: this.props.testID,
       style: [styles.pickerAndroid, this.props.style],
     };
 
     return <Picker ref={REF_PICKER} {...nativeProps} />;
-  },
-
-  _onFocusChange: function(event: Event) {
-    if (this.props.onFocus && event.nativeEvent.isFocused) {
-      this.props.onFocus();
-    } else if (this.props.onDismissed && !event.nativeEvent.isFocused) {
-      this.props.onDismissed();
-    }
   },
 
   _onChange: function(event: Event) {
@@ -116,15 +109,24 @@ var PickerAndroid = React.createClass({
         this.props.onValueChange(null, position);
       }
     }
+    this._lastNativePosition = event.nativeEvent.position;
+    this.forceUpdate();
+  },
 
+  componentDidMount: function() {
+    this._lastNativePosition = this.state.initialSelectedIndex;
+  },
+
+  componentDidUpdate: function() {
     // The picker is a controlled component. This means we expect the
     // on*Change handlers to be in charge of updating our
     // `selectedValue` prop. That way they can also
     // disallow/undo/mutate the selection of certain values. In other
     // words, the embedder of this component should be the source of
     // truth, not the native component.
-    if (this.refs[REF_PICKER] && this.state.selectedIndex != -1 && this.state.selectedIndex !== event.nativeEvent.position) {
+    if (this.refs[REF_PICKER] && this.state.selectedIndex !== this._lastNativePosition) {
       this.refs[REF_PICKER].setNativeProps({selected: this.state.selectedIndex});
+      this._lastNativePosition = this.state.selectedIndex;
     }
   },
 });
