@@ -11,6 +11,7 @@
  */
 'use strict';
 
+const NativeEventEmitter = require('NativeEventEmitter');
 const StatusBar = require('StatusBar');
 
 import type {StatusBarStyle, StatusBarAnimation} from 'StatusBar';
@@ -18,17 +19,17 @@ import type {StatusBarStyle, StatusBarAnimation} from 'StatusBar';
 /**
  * Deprecated. Use `StatusBar` instead.
  */
-const StatusBarIOS = {
+class StatusBarIOS extends NativeEventEmitter {
 
   setStyle(style: StatusBarStyle, animated?: boolean) {
     console.warn('`StatusBarIOS.setStyle` is deprecated. Use `StatusBar.setBarStyle` instead.');
     StatusBar.setBarStyle(style, animated);
-  },
+  }
 
   setHidden(hidden: boolean, animation?: StatusBarAnimation) {
     console.warn('`StatusBarIOS.setHidden` is deprecated. Use `StatusBar.setHidden` instead.');
     StatusBar.setHidden(hidden, animation);
-  },
+  }
 
   setNetworkActivityIndicatorVisible(visible: boolean) {
     console.warn(
@@ -36,7 +37,7 @@ const StatusBarIOS = {
       'Use `StatusBar.setNetworkActivityIndicatorVisible` instead.'
     );
     StatusBar.setNetworkActivityIndicatorVisible(visible);
-  },
+  }
 };
 
-module.exports = StatusBarIOS;
+module.exports = new StatusBarIOS('StatusBarManager');
