@@ -28,26 +28,50 @@ var TRACE_TAG_JSC_CALLS = 1 << 27;
 
 var _enabled = false;
 var _asyncCookie = 0;
-var _ReactPerf = null;
-function ReactPerf() {
-  if (!_ReactPerf) {
-    _ReactPerf = require('ReactPerf');
+var _ReactDebugTool = null;
+var _ReactComponentTreeDevtool = null;
+function ReactDebugTool() {
+  if (!_ReactDebugTool) {
+    _ReactDebugTool = require('ReactDebugTool');
   }
-  return _ReactPerf;
+  return _ReactDebugTool;
 }
+function ReactComponentTreeDevtool() {
+  if (!_ReactComponentTreeDevtool) {
+    _ReactComponentTreeDevtool = require('ReactComponentTreeDevtool');
+  }
+  return _ReactComponentTreeDevtool;
+}
+
+var ReactSystraceDevtool = {
+  onBeginReconcilerTimer(debugID, timerType) {
+    var displayName = ReactComponentTreeDevtool().getDisplayName(debugID);
+    Systrace.beginEvent(`ReactReconciler.${timerType}(${displayName})`);
+  },
+  onEndReconcilerTimer(debugID, timerType) {
+    Systrace.endEvent();
+  },
+  onBeginLifeCycleTimer(debugID, timerType) {
+    var displayName = ReactComponentTreeDevtool().getDisplayName(debugID);
+    Systrace.beginEvent(`${displayName}.${timerType}()`);
+  },
+  onEndLifeCycleTimer(debugID, timerType) {
+    Systrace.endEvent();
+  },
+};
 
 var Systrace = {
   setEnabled(enabled: boolean) {
     if (_enabled !== enabled) {
       if (enabled) {
         global.nativeTraceBeginLegacy && global.nativeTraceBeginLegacy(TRACE_TAG_JSC_CALLS);
+        ReactDebugTool().addDevtool(ReactSystraceDevtool);
       } else {
         global.nativeTraceEndLegacy && global.nativeTraceEndLegacy(TRACE_TAG_JSC_CALLS);
+        ReactDebugTool().removeDevtool(ReactSystraceDevtool);
       }
     }
     _enabled = enabled;
-
-    ReactPerf().enableMeasure = enabled;
   },
 
   /**
@@ -101,24 +125,6 @@ var Systrace = {
       global.nativeTraceCounter &&
         global.nativeTraceCounter(TRACE_TAG_REACT_APPS, profileName, value);
     }
-  },
-
-  reactPerfMeasure(objName: string, fnName: string, func: any): any {
-    return function (component) {
-      if (!_enabled) {
-        return func.apply(this, arguments);
-      }
-
-      var name = objName === 'ReactCompositeComponent' && this.getName() || '';
-      Systrace.beginEvent(`${objName}.${fnName}(${name})`);
-    var ret = func.apply(this, arguments);
-      Systrace.endEvent();
-      return ret;
-    };
-  },
-
-  swizzleReactPerf() {
-    ReactPerf().injection.injectMeasure(Systrace.reactPerfMeasure);
   },
 
   /**
