@@ -15,21 +15,10 @@ static NSDictionary *RCTParseKeyboardNotification(NSNotification *notification);
 
 @implementation RCTKeyboardObserver
 
-@synthesize bridge = _bridge;
-
 RCT_EXPORT_MODULE()
 
-- (instancetype)init
+- (void)startObserving
 {
-  // We're only overriding this to ensure the module gets created at startup
-  // TODO (t11106126): Remove once we have more declarative control over module setup.
-  return [super init];
-}
-
-- (void)setBridge:(RCTBridge *)bridge
-{
-  _bridge = bridge;
-
   NSNotificationCenter *nc = [NSNotificationCenter defaultCenter];
 
 #define ADD_KEYBOARD_HANDLER(NAME, SELECTOR) \
@@ -46,7 +35,17 @@ RCT_EXPORT_MODULE()
 
 }
 
-- (void)dealloc
+- (NSArray<NSString *> *)supportedEvents
+{
+  return @[@"keyboardWillShow",
+           @"keyboardDidShow",
+           @"keyboardWillHide",
+           @"keyboardDidHide",
+           @"keyboardWillChangeFrame",
+           @"keyboardDidChangeFrame"];
+}
+
+- (void)stopObserving
 {
   [[NSNotificationCenter defaultCenter] removeObserver:self];
 }
@@ -54,8 +53,7 @@ RCT_EXPORT_MODULE()
 #define IMPLEMENT_KEYBOARD_HANDLER(EVENT) \
 - (void)EVENT:(NSNotification *)notification \
 { \
-  [_bridge.eventDispatcher \
-    sendDeviceEventWithName:@#EVENT \
+  [self sendEventWithName:@#EVENT \
     body:RCTParseKeyboardNotification(notification)]; \
 }
 
