@@ -9,7 +9,7 @@
 
 #import <UIKit/UIKit.h>
 
-#import "RCTBridgeModule.h"
+#import "RCTEventEmitter.h"
 #import "RCTConvert.h"
 
 @interface RCTConvert (UIStatusBar)
@@ -19,6 +19,6 @@
 
 @end
 
-@interface RCTStatusBarManager : NSObject <RCTBridgeModule>
+@interface RCTStatusBarManager : RCTEventEmitter
 
 @end
