@@ -55,7 +55,9 @@
   UIView *testView = self;
   UIView *clipView = nil;
   CGRect clipRect = self.bounds;
-  while (testView) {
+  // We will only look for a clipping view up the view hierarchy until we hit the root view.
+  BOOL passedRootView = NO;
+  while (testView && !passedRootView) {
     if (testView.clipsToBounds) {
       if (clipView) {
         CGRect testRect = [clipView convertRect:clipRect toView:testView];
@@ -67,6 +69,9 @@
         clipView = testView;
         clipRect = [self convertRect:self.bounds toView:clipView];
       }
+    }
+    if ([testView isReactRootView]) {
+      passedRootView = YES;
     }
     testView = testView.superview;
   }
