@@ -27,6 +27,7 @@ public abstract class ReactActivity extends Activity implements DefaultHardwareB
       "Overlay permissions needs to be granted in order for react native apps to run in dev mode";
 
   private @Nullable ReactInstanceManager mReactInstanceManager;
+  private @Nullable ReactRootView mReactRootView;
   private LifecycleState mLifecycleState = LifecycleState.BEFORE_RESUME;
   private boolean mDoRefresh = false;
 
@@ -138,7 +139,7 @@ public abstract class ReactActivity extends Activity implements DefaultHardwareB
     }
 
     mReactInstanceManager = createReactInstanceManager();
-    ReactRootView mReactRootView = createRootView();
+    mReactRootView = createRootView();
     mReactRootView.startReactApplication(mReactInstanceManager, getMainComponentName(), getLaunchOptions());
     setContentView(mReactRootView);
   }
@@ -168,6 +169,9 @@ public abstract class ReactActivity extends Activity implements DefaultHardwareB
   @Override
   protected void onDestroy() {
     super.onDestroy();
+
+    mReactRootView.unmountReactApplication();
+    mReactRootView = null;
 
     if (mReactInstanceManager != null) {
       mReactInstanceManager.destroy();
