@@ -166,12 +166,17 @@ public class ReactRootView extends SizeMonitoringFrameLayout implements RootView
   }
 
   @Override
+  protected void onAttachedToWindow() {
+    super.onAttachedToWindow();
+    if (mIsAttachedToInstance) {
+      getViewTreeObserver().addOnGlobalLayoutListener(getKeyboardListener());
+    }
+  }
+
+  @Override
   protected void onDetachedFromWindow() {
     super.onDetachedFromWindow();
-
-    if (mReactInstanceManager != null && mIsAttachedToInstance) {
-      mReactInstanceManager.detachRootView(this);
-      mIsAttachedToInstance = false;
+    if (mIsAttachedToInstance) {
       getViewTreeObserver().removeOnGlobalLayoutListener(getKeyboardListener());
     }
   }
@@ -217,6 +222,19 @@ public class ReactRootView extends SizeMonitoringFrameLayout implements RootView
     }
   }
 
+  /**
+   * Unmount the react application at this root view, reclaiming any JS memory associated with that
+   * application. If {@link #startReactApplication} is called, this method must be called before the
+   * ReactRootView is garbage collected (typically in your Activity's onDestroy, or in your Fragment's
+   * onDestroyView).
+   */
+  public void unmountReactApplication() {
+    if (mReactInstanceManager != null && mIsAttachedToInstance) {
+      mReactInstanceManager.detachRootView(this);
+      mIsAttachedToInstance = false;
+    }
+  }
+
   /* package */ String getJSModuleName() {
     return Assertions.assertNotNull(mJSModuleName);
   }
@@ -250,6 +268,17 @@ public class ReactRootView extends SizeMonitoringFrameLayout implements RootView
     mIsAttachedToInstance = true;
     Assertions.assertNotNull(mReactInstanceManager).attachMeasuredRootView(this);
     getViewTreeObserver().addOnGlobalLayoutListener(getKeyboardListener());
+  }
+
+  @Override
+  protected void finalize() throws Throwable {
+    super.finalize();
+    Assertions.assertCondition(
+      !mIsAttachedToInstance,
+      "The application this ReactRootView was rendering was not unmounted before the ReactRootView " +
+        "was garbage collected. This usually means that your application is leaking large amounts of " +
+        "memory. To solve this, make sure to call ReactRootView#unmountReactApplication in the onDestroy() " +
+        "of your hosting Activity or in the onDestroyView() of your hosting Fragment.");
   }
 
   private class KeyboardListener implements ViewTreeObserver.OnGlobalLayoutListener {
