@@ -72,7 +72,7 @@ public class CatalystInstanceImpl implements CatalystInstance {
       final ReactQueueConfigurationSpec ReactQueueConfigurationSpec,
       final JavaScriptExecutor jsExecutor,
       final NativeModuleRegistry registry,
-      final JavaScriptModulesConfig jsModulesConfig,
+      final JavaScriptModuleRegistry jsModuleRegistry,
       final JSBundleLoader jsBundleLoader,
       NativeModuleCallExceptionHandler nativeModuleCallExceptionHandler) {
     FLog.d(ReactConstants.TAG, "Initializing React Bridge.");
@@ -81,7 +81,7 @@ public class CatalystInstanceImpl implements CatalystInstance {
         new NativeExceptionHandler());
     mBridgeIdleListeners = new CopyOnWriteArrayList<>();
     mJavaRegistry = registry;
-    mJSModuleRegistry = new JavaScriptModuleRegistry(CatalystInstanceImpl.this, jsModulesConfig);
+    mJSModuleRegistry = jsModuleRegistry;
     mJSBundleLoader = jsBundleLoader;
     mNativeModuleCallExceptionHandler = nativeModuleCallExceptionHandler;
     mTraceListener = new JSProfilerTraceListener();
@@ -93,7 +93,7 @@ public class CatalystInstanceImpl implements CatalystInstance {
             public ReactBridge call() throws Exception {
               Systrace.beginSection(Systrace.TRACE_TAG_REACT_JAVA_BRIDGE, "initializeBridge");
               try {
-                return initializeBridge(jsExecutor, jsModulesConfig);
+                return initializeBridge(jsExecutor);
               } finally {
                 Systrace.endSection(Systrace.TRACE_TAG_REACT_JAVA_BRIDGE);
               }
@@ -104,9 +104,7 @@ public class CatalystInstanceImpl implements CatalystInstance {
     }
   }
 
-  private ReactBridge initializeBridge(
-      JavaScriptExecutor jsExecutor,
-      JavaScriptModulesConfig jsModulesConfig) {
+  private ReactBridge initializeBridge(JavaScriptExecutor jsExecutor) {
     mReactQueueConfiguration.getJSQueueThread().assertIsOnThread();
     Assertions.assertCondition(mBridge == null, "initializeBridge should be called once");
 
@@ -126,7 +124,7 @@ public class CatalystInstanceImpl implements CatalystInstance {
     try {
       bridge.setGlobalVariable(
           "__fbBatchedBridgeConfig",
-          buildModulesConfigJSONProperty(mJavaRegistry, jsModulesConfig));
+          buildModulesConfigJSONProperty(mJavaRegistry));
       bridge.setGlobalVariable(
           "__RCTProfileIsProfiling",
           Systrace.isTracing(Systrace.TRACE_TAG_REACT_APPS) ? "true" : "false");
@@ -280,7 +278,8 @@ public class CatalystInstanceImpl implements CatalystInstance {
 
   @Override
   public <T extends JavaScriptModule> T getJSModule(ExecutorToken executorToken, Class<T> jsInterface) {
-    return Assertions.assertNotNull(mJSModuleRegistry).getJavaScriptModule(executorToken, jsInterface);
+    return Assertions.assertNotNull(mJSModuleRegistry)
+      .getJavaScriptModule(this, executorToken, jsInterface);
   }
 
   @Override
@@ -350,17 +349,13 @@ public class CatalystInstanceImpl implements CatalystInstance {
     mBridge.setGlobalVariable(propName, jsonValue);
   }
 
-  private String buildModulesConfigJSONProperty(
-      NativeModuleRegistry nativeModuleRegistry,
-      JavaScriptModulesConfig jsModulesConfig) {
+  private String buildModulesConfigJSONProperty(NativeModuleRegistry nativeModuleRegistry) {
     StringWriter stringWriter = new StringWriter();
     JsonWriter writer = new JsonWriter(stringWriter);
     try {
       writer.beginObject();
       writer.name("remoteModuleConfig");
       nativeModuleRegistry.writeModuleDescriptions(writer);
-      writer.name("localModulesConfig");
-      jsModulesConfig.writeModuleDescriptions(writer);
       writer.endObject();
       return stringWriter.toString();
     } catch (IOException ioe) {
@@ -504,7 +499,7 @@ public class CatalystInstanceImpl implements CatalystInstance {
     private @Nullable ReactQueueConfigurationSpec mReactQueueConfigurationSpec;
     private @Nullable JSBundleLoader mJSBundleLoader;
     private @Nullable NativeModuleRegistry mRegistry;
-    private @Nullable JavaScriptModulesConfig mJSModulesConfig;
+    private @Nullable JavaScriptModuleRegistry mJSModuleRegistry;
     private @Nullable JavaScriptExecutor mJSExecutor;
     private @Nullable NativeModuleCallExceptionHandler mNativeModuleCallExceptionHandler;
 
@@ -519,8 +514,8 @@ public class CatalystInstanceImpl implements CatalystInstance {
       return this;
     }
 
-    public Builder setJSModulesConfig(JavaScriptModulesConfig jsModulesConfig) {
-      mJSModulesConfig = jsModulesConfig;
+    public Builder setJSModuleRegistry(JavaScriptModuleRegistry jsModuleRegistry) {
+      mJSModuleRegistry = jsModuleRegistry;
       return this;
     }
 
@@ -545,7 +540,7 @@ public class CatalystInstanceImpl implements CatalystInstance {
           Assertions.assertNotNull(mReactQueueConfigurationSpec),
           Assertions.assertNotNull(mJSExecutor),
           Assertions.assertNotNull(mRegistry),
-          Assertions.assertNotNull(mJSModulesConfig),
+          Assertions.assertNotNull(mJSModuleRegistry),
           Assertions.assertNotNull(mJSBundleLoader),
           Assertions.assertNotNull(mNativeModuleCallExceptionHandler));
     }
