@@ -131,7 +131,7 @@ public class MessageQueueThreadImpl implements MessageQueueThread {
       case MAIN_UI:
         return createForMainThread(spec.getName(), exceptionHandler);
       case NEW_BACKGROUND:
-        return startNewBackgroundThread(spec.getName(), exceptionHandler);
+        return startNewBackgroundThread(spec.getName(), spec.getStackSize(), exceptionHandler);
       default:
         throw new RuntimeException("Unknown thread type: " + spec.getThreadType());
     }
@@ -165,17 +165,27 @@ public class MessageQueueThreadImpl implements MessageQueueThread {
     return mqt;
   }
 
-  /**
-   * Creates  and starts a new MessageQueueThreadImpl encapsulating a new Thread with a new Looper
-   * running on it. Give it a name for easier debugging. When this method exits, the new
-   * MessageQueueThreadImpl is ready to receive events.
-   */
   public static MessageQueueThreadImpl startNewBackgroundThread(
       final String name,
       QueueThreadExceptionHandler exceptionHandler) {
+    return startNewBackgroundThread(
+        name,
+        MessageQueueThreadSpec.DEFAULT_STACK_SIZE_BYTES,
+        exceptionHandler);
+  }
+
+  /**
+   * Creates and starts a new MessageQueueThreadImpl encapsulating a new Thread with a new Looper
+   * running on it. Give it a name for easier debugging and optionally a suggested stack size.
+   * When this method exits, the new MessageQueueThreadImpl is ready to receive events.
+   */
+  public static MessageQueueThreadImpl startNewBackgroundThread(
+      final String name,
+      long stackSize,
+      QueueThreadExceptionHandler exceptionHandler) {
     final SimpleSettableFuture<Looper> looperFuture = new SimpleSettableFuture<>();
     final SimpleSettableFuture<MessageQueueThread> mqtFuture = new SimpleSettableFuture<>();
-    Thread bgThread = new Thread(
+    Thread bgThread = new Thread(null,
         new Runnable() {
           @Override
           public void run() {
@@ -186,7 +196,7 @@ public class MessageQueueThreadImpl implements MessageQueueThread {
 
             Looper.loop();
           }
-        }, "mqt_" + name);
+        }, "mqt_" + name, stackSize);
     bgThread.start();
 
     Looper myLooper = looperFuture.getOrThrow();
