@@ -9,6 +9,8 @@
 
 package com.facebook.react.bridge.queue;
 
+import android.os.Build;
+
 import javax.annotation.Nullable;
 
 import com.facebook.infer.annotation.Assertions;
@@ -20,6 +22,8 @@ import com.facebook.infer.annotation.Assertions;
  * specs allows the Threads to be created fully configured.
  */
 public class ReactQueueConfigurationSpec {
+
+  private static final long LEGACY_STACK_SIZE_BYTES = 2000000;
 
   private final MessageQueueThreadSpec mNativeModulesQueueThreadSpec;
   private final MessageQueueThreadSpec mJSQueueThreadSpec;
@@ -44,10 +48,12 @@ public class ReactQueueConfigurationSpec {
   }
 
   public static ReactQueueConfigurationSpec createDefault() {
+    MessageQueueThreadSpec spec = Build.VERSION.SDK_INT < 21 ?
+        MessageQueueThreadSpec.newBackgroundThreadSpec("native_modules", LEGACY_STACK_SIZE_BYTES) :
+        MessageQueueThreadSpec.newBackgroundThreadSpec("native_modules");
     return builder()
         .setJSQueueThreadSpec(MessageQueueThreadSpec.newBackgroundThreadSpec("js"))
-        .setNativeModulesQueueThreadSpec(
-            MessageQueueThreadSpec.newBackgroundThreadSpec("native_modules"))
+        .setNativeModulesQueueThreadSpec(spec)
         .build();
   }
 
