@@ -95,10 +95,14 @@ import com.facebook.csslayout.Spacing;
 
   @Override
   public void draw(Canvas canvas) {
-    if ((!CSSConstants.isUndefined(mBorderRadius) && mBorderRadius > 0) || mBorderCornerRadii != null) {
-      drawRoundedBackgroundWithBorders(canvas);
-    } else {
+    updatePathEffect();
+    boolean roundedBorders = mBorderCornerRadii != null ||
+        (!CSSConstants.isUndefined(mBorderRadius) && mBorderRadius > 0);
+
+    if ((mBorderStyle == null || mBorderStyle == BorderStyle.SOLID) && !roundedBorders) {
       drawRectangularBackgroundWithBorders(canvas);
+    } else {
+      drawRoundedBackgroundWithBorders(canvas);
     }
   }
 
@@ -231,7 +235,6 @@ import com.facebook.csslayout.Spacing;
       mPaint.setColor(ColorUtil.multiplyColorAlpha(borderColor, mAlpha));
       mPaint.setStyle(Paint.Style.STROKE);
       mPaint.setStrokeWidth(fullBorderWidth);
-      mPaint.setPathEffect(mPathEffectForBorderStyle);
       canvas.drawPath(mPathForBorderRadius, mPaint);
     }
   }
@@ -298,10 +301,17 @@ import com.facebook.csslayout.Spacing;
         bottomLeftRadius + extraRadiusForOutline
       },
       Path.Direction.CW);
+  }
 
+  /**
+   * Set type of border
+   */
+  private void updatePathEffect() {
     mPathEffectForBorderStyle = mBorderStyle != null
         ? mBorderStyle.getPathEffect(getFullBorderWidth())
         : null;
+
+    mPaint.setPathEffect(mPathEffectForBorderStyle);
   }
 
   /**
