@@ -1,16 +1,15 @@
 /**
- * Copyright (c) 2015-present, Facebook, Inc.
+ * Copyright (c) 2016-present, Facebook, Inc.
  * All rights reserved.
  *
  * This source code is licensed under the BSD-style license found in the
  * LICENSE file in the root directory of this source tree. An additional grant
  * of patent rights can be found in the PATENTS file in the same directory.
+ *
+ * @flow
  */
+'use strict';
 
-#import <Foundation/Foundation.h>
-
-#import "RCTBridgeModule.h"
-
-@interface RCTKeyboardObserver : NSObject <RCTBridgeModule>
-
-@end
+// This is a forwarding module to allow React to require React Native internals
+// as node dependency
+module.exports = require('InitializeJavaScriptAppEngine');
